requirements.txt
flask
opencv-python
numpy
requests
gunicorn
python-telegram-bot>=20.0
